renames Vimeo-Video to match class name pattern
